refactor: moved method
