+++ anonmodel/document.go
@@ -1,31 +1,10 @@
 package anonmodel
 
-import (
-	"fmt"
-	"strings"
-)
-
 // Document represents a data object of any type uploaded by the client
 type Document map[string]interface{}
 
 // Documents represents an array of data objects of any type uploaded by the client
 type Documents []Document
-
-func validateFieldName(field string) error {
-	if field == "_id" {
-		return ErrValidation("Validation error: the '_id' field is not allowed")
-	}
-
-	if strings.HasPrefix(field, "__") {
-		return ErrValidation(fmt.Sprintf("Validation error (%v): document fields starting with '__' are reserved by the anonymization server", field))
-	}
-
-	if strings.ContainsAny(field, ".$") {
-		return ErrValidation(fmt.Sprintf("Validation error (%v): document fields containing either '.' or '$' are not allowed", field))
-	}
-
-	return nil
-}
 
 func (document Document) validate() error {
 	for key := range document {
